chore: удаление избыточного роутера
